+++ src/hooks/useTransfer.js
@@ -12,7 +12,7 @@ import {ActionType, TransactionStatus} from '../enums';
 import {starknet} from '../libs';
 import {useL1Token, useTokens} from '../providers/TokensProvider';
 import {useAmount, useSelectedToken} from '../providers/TransferProvider';
-import {useTransfers} from '../providers/TransfersProvider';
+import {useTransfersLog} from '../providers/TransfersLogProvider';
 import {useL1Wallet, useL2Wallet} from '../providers/WalletsProvider';
 import utils from '../utils';
 import {useL1TokenBridgeContract, useTokenBridgeContract, useTokenContract} from './useContract';
@@ -241,7 +241,7 @@ const useTransfer = () => {
   const showErrorModal = useErrorModal();
   const hideModal = useHideModal();
   const showTransactionSubmittedModal = useTransactionSubmittedModal();
-  const {addTransfer} = useTransfers();
+  const {addTransfer} = useTransfersLog();
   const {updateTokenBalance} = useTokens();
   const [, , clearAmount] = useAmount();
