+++ src/providers/TransfersLogProvider/TransfersLogProvider.js
@@ -8,13 +8,13 @@ import {useLogger} from '../../hooks';
 import {starknet} from '../../libs';
 import utils from '../../utils';
 import {useBlockHash} from '../BlockHashProvider';
-import {TransfersContext} from './transfers-context';
-import {actions, initialState, reducer} from './transfers-reducer';
+import {TransfersLogContext} from './transfers-log-context';
+import {actions, initialState, reducer} from './transfers-log-reducer';
 
 const {LOCAL_STORAGE_TRANSFERS_KEY} = constants;
 
-export const TransfersProvider = ({children}) => {
-  const logger = useLogger(TransfersProvider.displayName);
+export const TransfersLogProvider = ({children}) => {
+  const logger = useLogger(TransfersLogProvider.displayName);
   const [transfers, dispatch] = useReducer(reducer, initialState);
   const blockHash = useBlockHash();
 
@@ -89,11 +89,11 @@ export const TransfersProvider = ({children}) => {
     addTransfer
   };
 
-  return <TransfersContext.Provider value={context}>{children}</TransfersContext.Provider>;
+  return <TransfersLogContext.Provider value={context}>{children}</TransfersLogContext.Provider>;
 };
 
-TransfersProvider.displayName = 'TransfersProvider';
+TransfersLogProvider.displayName = 'TransfersLogProvider';
 
-TransfersProvider.propTypes = {
+TransfersLogProvider.propTypes = {
   children: PropTypes.oneOfType([PropTypes.object, PropTypes.array])
 };
