+++ src/providers/TransfersLogProvider/transfers-log-reducer.js
