--- src/providers/TransfersProvider/index.js
@@ -1,4 +1,0 @@
-export * from './transfers-reducer';
-export * from './transfers-context';
-export * from './transfers-hooks';
-export * from './TransfersProvider';
