+++ src/providers/index.js
@@ -2,7 +2,7 @@ import {BlockHashProvider} from './BlockHashProvider';
 import {MenuProvider} from './MenuProvider';
 import {StoreProvider} from './StoreProvider/StoreProvider';
 import {TransferProvider} from './TransferProvider';
-import {TransfersProvider} from './TransfersProvider';
+import {TransfersLogProvider} from './TransfersLogProvider';
 import {WalletProvider} from './WalletProvider/WalletProvider';
 import {WalletsProvider} from './WalletsProvider';
 import {combineProviders} from './combine-providers';
@@ -14,5 +14,5 @@ export const Providers = combineProviders([
   WalletProvider,
   WalletsProvider,
   BlockHashProvider,
-  TransfersProvider
+  TransfersLogProvider
 ]);
